Updates for video overlay, animated stats, css
Style sheet isn't picking up some elements added on the fly, will debug
later. beatPregnant.tsv is a tab-separated file with the stats that D3
reads. Easy to add other countries' stats as found. Added youtube
overlay on the SVG to play "one billion rising" video -- a generic or
country-specific.
Lots of wasted code and lack of optimization. Really need to clean this
up, a little embarrassed by it, but it's getting the job done. Would be
helpful to lookup some D3 calls and learn D3 better rather than rely on
jQuery.